Create IP2Location.py
A Python script that takes an IP address as input and provides the geolocation of that IP address. Includes latitude and longitude, Country Code, Country Name, and Region. Leverages the IP2Location API. You will need to register for an API Key at IP2Location.com and then update the python script with your API.